get glitchtip projects from namespaces instead of apps
