+++ reconcile/gql_definitions/glitchtip/glitchtip_project_fragment.py
@@ -1,0 +1,93 @@
+"""
+Generated by qenerate plugin=pydantic_v1. DO NOT MODIFY MANUALLY!
+"""
+from enum import Enum  # noqa: F401 # pylint: disable=W0611
+from typing import (  # noqa: F401 # pylint: disable=W0611
+    Any,
+    Callable,
+    Optional,
+    Union,
+)
+
+from pydantic import (  # noqa: F401 # pylint: disable=W0611
+    BaseModel,
+    Extra,
+    Field,
+    Json,
+)
+
+
+class GlitchtipOrganizationV1(BaseModel):
+    name: str = Field(..., alias="name")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
+
+
+class GlitchtipRoleV1(BaseModel):
+    organization: GlitchtipOrganizationV1 = Field(..., alias="organization")
+    role: str = Field(..., alias="role")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
+
+
+class UserV1(BaseModel):
+    org_username: str = Field(..., alias="org_username")
+    github_username: str = Field(..., alias="github_username")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
+
+
+class RoleV1(BaseModel):
+    glitchtip_roles: Optional[list[GlitchtipRoleV1]] = Field(
+        ..., alias="glitchtip_roles"
+    )
+    users: list[UserV1] = Field(..., alias="users")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
+
+
+class GlitchtipTeamV1(BaseModel):
+    name: str = Field(..., alias="name")
+    roles: list[RoleV1] = Field(..., alias="roles")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
+
+
+class GlitchtipInstanceV1(BaseModel):
+    name: str = Field(..., alias="name")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
+
+
+class GlitchtipProjectsV1_GlitchtipOrganizationV1(BaseModel):
+    name: str = Field(..., alias="name")
+    instance: GlitchtipInstanceV1 = Field(..., alias="instance")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
+
+
+class GlitchtipProject(BaseModel):
+    name: str = Field(..., alias="name")
+    platform: str = Field(..., alias="platform")
+    teams: list[GlitchtipTeamV1] = Field(..., alias="teams")
+    organization: GlitchtipProjectsV1_GlitchtipOrganizationV1 = Field(
+        ..., alias="organization"
+    )
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
